# contracts/openapi/servicehub-v1.yaml
openapi: 3.1.0
info:
  title: ServiceHub 工单平台 API
  version: 1.0.0-draft.1
  description: |
    Vue 3 门户与 Spring Boot 单体服务的共享 v1 契约。所有业务 API 都由服务端执行对象级、流程动作和数据范围鉴权；UI 隐藏不是授权控制。
    附件与知识库文档必须先进入隔离区，完成服务端白名单校验和恶意文件扫描后才可下载、索引或发布。消息仅通过受控站内信和受管通道适配器投递；不得以通用回调、任意 URL 或客户端指定收件人替代。
    本版本不包含 AI、向量检索、自动生成、自动问答或自动处置能力。
    外部监控、CMDB、日志与 APM 仅通过受管连接器集成；本版本不实现自建监控平台、eBPF、自动发现 CMDB 或数据库运维平台。
servers:
  - url: /api/v1
    description: 同源网关路径
...
paths:
...
  /tickets/{ticketId}/attachments:
    get:
      tags: [Attachments]
      summary: 查询工单附件元数据
      description: |
        仅返回当前主体有权查看的工单附件元数据，不返回对象存储路径、扫描报告、原始文件名之外的敏感载荷或下载地址。
        列表与每个附件均须逐对象授权；隔离、扫描中、拒绝和逻辑删除保留的附件不提供下载能力。
      operationId: listTicketAttachments
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 已授权工单的附件元数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketAttachmentPage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Attachments]
      summary: 上传工单附件至隔离区
      description: |
        文件名、扩展名、Content-Type、路径和客户端声明的 MIME 类型均不可信。服务端以文件签名和内容复核允许类型、大小、数量和用途，
        生成随机对象键后写入隔离区并异步扫描。上传成功只表示已受控接收，绝不表示可下载、可展示或可对外投递。
      operationId: uploadTicketAttachment
      x-replay-protection: subject-ticket-attachment-request-fingerprint
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TicketAttachmentUploadRequest'
      responses:
        '202':
          description: 文件已进入隔离和扫描队列；扫描通过前不能下载或被知识库发布引用。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentUploadAccepted'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/AttachmentRejected'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /admin/backoffice-access/{iamUserId}:
    get:
      tags: [BackofficeAccessAdmin]
      summary: 查询 IAM 用户的后台授权记录
      description: |
        仅平台管理员可调用。路径 IAM ID 只能定位已同步且启用的 IAM 用户；返回身份只读摘要及本地后台启停、角色、数据范围与版本。
        未配置后台授权不代表 IAM 用户不可登录前台，而是只具有普通提单人能力。不得通过此接口编辑 IAM 主数据或枚举人员目录。
      operationId: getBackofficeAccess
      x-required-platform-roles: [PLATFORM_ADMIN]
      parameters:
        - name: iamUserId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9._:-]{1,128}$'
      responses:
        '200':
          description: 已校验 IAM 身份的后台授权摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackofficeAccessResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
    put:
      tags: [BackofficeAccessAdmin]
      summary: 替换 IAM 用户的后台角色与数据范围
      description: |
        仅平台管理员可调用。服务端以已认证管理员身份审计，并校验 IAM 用户状态、乐观锁、角色白名单、数据范围格式、禁止自我提权和最后平台管理员保护。
        请求不得提交或修改密码、令牌、姓名、登录名、组织、岗位或任意 IAM 主数据；IAM ID 只是已验证 SSO 主体的绑定键，不能作为登录凭据。
      operationId: replaceBackofficeAccess
      x-required-platform-roles: [PLATFORM_ADMIN]
      parameters:
        - name: iamUserId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9._:-]{1,128}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackofficeAccessUpdateRequest'
      responses:
        '200':
          description: 已持久化且已审计的后台授权版本
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackofficeAccessResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /tickets/{ticketId}/configuration-items:
    get:
      tags: [Cmdb]
      summary: 查询工单已关联的配置项
      description: 先校验工单对象权限，再逐条校验 CI 当前可见性；不能因已存在历史关联而泄露调岗后无权的 CI 详情。
      operationId: listTicketConfigurationItemAssociations
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 已授权工单的可见 CI 关联
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketConfigurationItemAssociationPage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Cmdb]
      summary: 关联已授权 CMDB 配置项至工单
      description: 服务端同时验证工单动作权限、CI 当前可见性、服务/组织范围、关联策略及重复关系；客户端不能设置 CI 详情、影响范围、CMDB 来源或关联审批结论。
      operationId: createTicketConfigurationItemAssociation
      x-replay-protection: subject-ticket-configuration-item-association-request-fingerprint
      parameters:
        - $ref: '#/components/parameters/TicketId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketConfigurationItemAssociationRequest'
      responses:
        '201':
          description: CI 关联已创建并审计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketConfigurationItemAssociation'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SERVICEHUB_SESSION
      description: 由 Spring Boot 维护的 HttpOnly、Secure、SameSite 会话 Cookie；浏览器不得将访问令牌写入 localStorage。
    integrationCallbackSignature:
      type: apiKey
      in: header
      name: X-Integration-Signature
      description: 仅受管连接器回调使用的签名头。服务端必须基于受管 secretRef、原始请求体、时间戳和事件 ID 验签；该头不是用户会话或通用 API 密钥。
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: 对所有非安全方法必传；令牌绑定当前会话。
      schema:
        type: string
        minLength: 20
        maxLength: 512
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        客户端为一次写入意图生成的 UUID。服务端将键绑定到认证主体、接口、目标工单/任务和规范化请求指纹；
        同一绑定在受控保留期内仅执行一次并返回原结果，使用相同键但不同指纹必须返回 IDEMPOTENCY_CONFLICT。
        键不可跨主体、工单、任务或操作复用，且服务端记录首次接收时间、结果和审计关联以防重放。
      schema:
        type: string
        format: uuid
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: string
        pattern: '^TKT-[0-9]{8}-[0-9]{6}$'
...
  responses:
    ValidationFailed:
      description: 请求字段或服务目录规则校验失败
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthenticated:
      description: 会话不存在、过期或 IAM 身份不可用
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: 已认证但不具备当前对象、动作或数据范围权限
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 对象不存在或按防枚举策略不可见
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 幂等键冲突、状态冲突或乐观锁冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
    AttachmentRejected:
      description: 文件超过受控限制、类型不在服务端白名单内或文件内容校验失败；不返回规则细节或恶意样本信息。
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    RateLimited:
      description: 请求频率超过当前主体或来源限制
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    PlatformRole:
      type: string
      enum: [REQUESTER, FIRST_LINE_SUPPORT, SECOND_LINE_SUPPORT, APPROVER, SERVICE_MANAGER, SLA_MANAGER, REPORT_VIEWER, PLATFORM_ADMIN, AUDITOR]
...
    ConfigurationItemSummary:
      type: object
      additionalProperties: false
      required: [id, name, type]
      properties:
        id:
          type: string
          pattern: '^CI-[A-Za-z0-9_-]{8,64}$'
        name:
          type: string
          maxLength: 200
        type:
          type: string
          maxLength: 64
...
    TicketConfigurationItemAssociationRequest:
      type: object
      additionalProperties: false
      required: [configurationItemId, version, reason]
      properties:
        configurationItemId:
          type: string
          pattern: '^CI-[A-Za-z0-9_-]{8,64}$'
        version:
          type: integer
          minimum: 0
          description: 工单乐观锁版本。
        reason:
          type: string
          minLength: 1
          maxLength: 500
    TicketConfigurationItemAssociation:
      type: object
      additionalProperties: false
      required: [configurationItem, associatedAt, associatedBy, auditEventId]
      properties:
        configurationItem:
          $ref: '#/components/schemas/ConfigurationItemSummary'
        associatedAt:
          type: string
          format: date-time
        associatedBy:
          $ref: '#/components/schemas/IdentitySnapshot'
        auditEventId:
          type: string
          maxLength: 128
    TicketConfigurationItemAssociationPage:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/TicketConfigurationItemAssociation'
        total:
          type: integer
          minimum: 0
...
    AttachmentPurpose:
      type: string
      enum: [TICKET_EVIDENCE, USER_FEEDBACK, HANDOVER_CHECKLIST, OTHER]
    AttachmentScanState:
      type: string
      description: 扫描状态由服务端及受管防病毒服务写入；客户端不能设置或覆盖。
      enum: [RECEIVED, SCANNING, SCAN_PASSED, QUARANTINED, REJECTED, SCAN_UNAVAILABLE]
    AttachmentRetentionState:
      type: string
      description: 删除为逻辑删除保留，不执行普通 API 物理删除；保留期和法务留存策略由服务端决定。
      enum: [ACTIVE, DELETE_REQUESTED, RETAINED]
    TicketAttachmentUploadRequest:
      type: object
      additionalProperties: false
      required: [file, purpose]
      x-server-enforced:
        - ticket-object-authorization
        - server-side-content-signature-and-allowlist-validation
        - per-ticket-file-count-and-size-limits
        - quarantine-storage-with-random-object-key
        - malware-scan-before-release
        - immutable-audit-event
      properties:
        file:
          type: string
          format: binary
          description: 文件名、扩展名、客户端 Content-Type 和路径不可信；服务端只接受受控类型、大小与内容签名一致的文件。
        purpose:
          $ref: '#/components/schemas/AttachmentPurpose'
        comment:
          type: string
          maxLength: 300
          description: 可选纯文本说明；服务端净化，不作为文件路径、MIME 类型或下载名。
    TicketAttachment:
      type: object
      additionalProperties: false
      required: [id, displayFileName, detectedMediaType, sizeBytes, purpose, scanState, retentionState, uploadedAt]
      properties:
        id:
          type: string
          pattern: '^ATT-[A-Za-z0-9_-]{8,64}$'
        displayFileName:
          type: string
          maxLength: 180
          description: 服务端生成并净化后的展示名；不含路径、控制字符或可执行内容。
        detectedMediaType:
          type: string
          maxLength: 100
          description: 服务端探测的受控 MIME 类型，不采信客户端声明。
        sizeBytes:
          type: integer
          minimum: 0
          maximum: 104857600
        purpose:
          $ref: '#/components/schemas/AttachmentPurpose'
        scanState:
          $ref: '#/components/schemas/AttachmentScanState'
        retentionState:
          $ref: '#/components/schemas/AttachmentRetentionState'
        downloadable:
          type: boolean
          description: 仅服务端判定为 SCAN_PASSED、未逻辑删除且当前主体仍具备下载权时为 true。
        uploadedAt:
          type: string
          format: date-time
        uploader:
          $ref: '#/components/schemas/IdentitySnapshot'
    TicketAttachmentPage:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/TicketAttachment'
        total:
          type: integer
          minimum: 0
    AttachmentUploadAccepted:
      type: object
      additionalProperties: false
      required: [attachment, auditEventId]
      properties:
        attachment:
          $ref: '#/components/schemas/TicketAttachment'
        auditEventId:
          type: string
          maxLength: 128
...
    IdentitySnapshot:
      type: object
      additionalProperties: false
      description: |
        工单事件、审批、分派和审计在发生时固化的身份事实。它不是当前 IAM 投影的外键视图：人员调岗、改名、停用或组织重组后，
        历史记录仍按本快照展示。写入仅由服务端从已校验的 IAM 投影生成，客户端不得提交或修改。
      required: [iamUserId, displayName, organizationName, projectionVersion, capturedAt]
      properties:
        iamUserId:
          type: string
          maxLength: 128
        displayName:
          type: string
          maxLength: 100
        organizationName:
          type: string
          maxLength: 200
        iamOrganizationId:
          type: string
          maxLength: 128
        positionName:
          type: string
          maxLength: 200
        iamPositionId:
          type: string
          maxLength: 128
        platformRoles:
          type: array
          maxItems: 20
          description: 事件发生时经服务端计算的角色摘要；仅在该历史记录的可见范围内返回。
          items:
            $ref: '#/components/schemas/PlatformRole'
        projectionVersion:
          type: string
          maxLength: 128
          description: 生成快照所依据的 IAM 投影来源版本/同步批次，用于历史一致性追溯。
        capturedAt:
          type: string
          format: date-time
...
    BackofficeAccessUpdateRequest:
      type: object
      additionalProperties: false
      required: [enabled, roleCodes, dataScopes, expectedVersion]
      properties:
        enabled:
          type: boolean
          description: 关闭后仅撤销后台能力，不影响普通用户通过 IAM 使用前台建单。
        roleCodes:
          type: array
          maxItems: 16
          uniqueItems: true
          items:
            type: string
            enum: [ROLE_FIRST_LINE_SUPPORT, ROLE_SECOND_LINE_SUPPORT, ROLE_APPROVER, ROLE_SERVICE_MANAGER, ROLE_SLA_MANAGER, ROLE_AUDITOR, ROLE_PLATFORM_ADMIN]
        dataScopes:
          type: array
          maxItems: 200
          uniqueItems: true
          items:
            type: object
            additionalProperties: false
            required: [scopeType, scopeId]
            properties:
              scopeType:
                type: string
                enum: [ORGANIZATION, SERVICE, QUEUE, CONFIGURATION_ITEM]
              scopeId:
                type: string
                maxLength: 128
                pattern: '^[A-Za-z0-9._:-]{1,128}$'
        expectedVersion:
          type: integer
          minimum: 0
          description: 后台授权记录乐观锁版本；首次配置为 0。
    BackofficeAccessResponse:
      type: object
      additionalProperties: false
      required: [user, access]
      properties:
        user:
          type: object
          additionalProperties: false
          required: [iamUserId, loginName, displayName, organizationIamId, organizationName]
          properties:
            iamUserId:
              type: string
            loginName:
              type: string
            displayName:
              type: string
            organizationIamId:
              type: string
            organizationName:
              type: string
        access:
          type: object
          additionalProperties: false
          required: [enabled, roleCodes, dataScopes, version]
          properties:
            enabled:
              type: boolean
            roleCodes:
              type: array
              items:
                type: string
            dataScopes:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [scopeType, scopeId]
                properties:
                  scopeType:
                    type: string
                  scopeId:
                    type: string
            version:
              type: integer
            updatedAt:
              type: [string, 'null']
              format: date-time
    ApiError:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
          description: 稳定问题类型，例如 /problems/validation-failed。
        title:
          type: string
          maxLength: 160
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          enum: [VALIDATION_FAILED, DYNAMIC_FORM_VALIDATION_FAILED, CATALOG_NOT_AVAILABLE, RULE_MATCH_NOT_ALLOWED, ATTACHMENT_REJECTED, ATTACHMENT_NOT_AVAILABLE, ATTACHMENT_SCAN_UNAVAILABLE, KNOWLEDGE_IMPORT_REJECTED, KNOWLEDGE_DOCUMENT_NOT_AVAILABLE, KNOWLEDGE_PUBLICATION_NOT_ALLOWED, INTEGRATION_CONFIGURATION_NOT_ALLOWED, INTEGRATION_CALLBACK_REJECTED, INTEGRATION_REPLAY_REJECTED, CONFIGURATION_ITEM_NOT_AVAILABLE, ALERT_NOT_AVAILABLE, OBSERVABILITY_LINK_NOT_AVAILABLE, NOTIFICATION_NOT_AVAILABLE, NOTIFICATION_RECIPIENT_NOT_ALLOWED, NOTIFICATION_TEMPLATE_REJECTED, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, VERSION_CONFLICT, IDEMPOTENCY_CONFLICT, INVALID_WORKFLOW_ACTION, INVALID_PROCESS_NODE, APPROVAL_REQUIRED, RATE_LIMITED, SERVICE_UNAVAILABLE, INTERNAL_ERROR]
        detail:
          type: string
          maxLength: 500
          description: 面向用户的安全摘要；禁止包含 SQL、堆栈、令牌、权限规则或对象敏感信息。
        requestId:
          type: string
          maxLength: 64
        fieldErrors:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field:
                type: string
                maxLength: 200
              code:
                type: string
                maxLength: 80
              message:
                type: string
                maxLength: 300
security:
  - sessionCookie: []